+++ docs/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* :feature:`7383` rotki will now update airdrops' data remotely without needing to update the app.
 * :bug:`7435` Fix the issue where, in rare cases, users receive a JavaScript error popup when minimizing the window.
 * :feature:`7357` rotki will now properly detect and decode Shutter token airdrop and delegation.
 * :feature:`-` Transactions contributing to Juicebox campaigns, such as the Free Alexey and Roman from Tornado, are now properly decoded.